+++ pymadvr/consts.py
