+++ pymadvr/errors.py
